# [JRuby] Optimize `scan()`: Remove duplicate `if (restLen() < patternsize()) return context.nil;` checks in `!headonly`.