@kou 2024-10-19
Thanks.

It's better that we compare performance change between `string`/`string_var` instead of `string`/`regexp` with/without this change.